+++ geolocator_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+
+- Add support to force using the Android Location Manager instead of the Android FusedLocationProvider.
+
 ## 1.0.0
 
 - Initial open-source release.